+++ packages/cli/docs-resolver/vitest.config.ts
@@ -15,8 +15,10 @@ const include = [
     "src/__test__/product-landing-page.test.ts",
     "src/__test__/versioned-root-landing-page.test.ts",
     "src/__test__/library-hardfail.test.ts",
+    "src/__test__/direct-api-spec.test.ts",
     "src/__test__/graphql-type-navigation.test.ts",
-    "src/__test__/convertThemeTabs.test.ts"
+    "src/__test__/convertThemeTabs.test.ts",
+    "src/__test__/root-changelog.test.ts"
 ];
 
 export default defineConfig({